% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finRes-package.R
\docType{package}
\name{finRes-package}
\alias{finRes}
\alias{finRes-package}
\title{finRes: Data-science and research in finance and financial economics in R made easy}
\description{
finRes is a set of packages that work in harmony because they share common data 
  representations and 'API' design. It is meant to facilitate data-science and research in finance and 
  financial economics in R. This package is designed to make it easy to install and load multiple
  'finRes' packages in a single step.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bautheac.github.io/finRes/}
  \item Report bugs at \url{https://github.com/bautheac/finRes/issues}
}

}
\author{
\strong{Maintainer}: Olivier Bauthéac \email{o.bautheac@gmail.com}

}
\keyword{internal}
