% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{finRes_packages}
\alias{finRes_packages}
\title{List all packages in the finRes suite.}
\usage{
finRes_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include finRes in the list?}
}
\description{
List all packages in the finRes suite.
}
\examples{
finRes_packages()
}
