% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{get_active_contract_tickers}
\alias{get_active_contract_tickers}
\alias{get_active_contract_tickers,FuturesHistorical-method}
\alias{get_active_contract_tickers,FuturesHistorical}
\title{Accesses the \code{active_contract_tickers} slot \linkS4class{FuturesMarket} S4 objects
  output of \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}.}
\usage{
get_active_contract_tickers(object)

\S4method{get_active_contract_tickers}{FuturesHistorical}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesMarket} or childs.}
}
\value{
A \link[data.table]{data.table} showing the futures active contract
  Bloomberg tickers for which the \code{object} provided contains data.
}
\description{
Access method for the \code{active_contract_tickers} slot of
  \linkS4class{FuturesMarket} S4 objects from the
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.
}
