% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{get_cftc_tickers}
\alias{get_cftc_tickers}
\alias{get_cftc_tickers,FuturesCFTC-method}
\alias{get_cftc_tickers,FuturesCFTC}
\title{Accesses the \code{cftc_tickers} slot \linkS4class{FuturesCFTC} S4 objects
  output of \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}.}
\usage{
get_cftc_tickers(object)

\S4method{get_cftc_tickers}{FuturesCFTC}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesCFTC}.}
}
\value{
A \link[data.table]{data.table} showing the futures position
  Bloomberg tickers for which the \code{object} provided contains data.
}
\description{
Access method for the \code{cftc_tickers} slot of \linkS4class{FuturesCFTC}
  S4 objects from the \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.
}
