% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{get_fields}
\alias{get_fields}
\alias{get_fields,DataInfo-method}
\alias{get_fields,DataInfo}
\alias{get_fields,DataHistorical-method}
\alias{get_fields,DataHistorical}
\title{Accesses the \code{fields} slot of various S4 objects output of
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}.}
\usage{
get_fields(object)

\S4method{get_fields}{DataInfo}(object)

\S4method{get_fields}{DataHistorical}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{DataInfo} or
\linkS4class{DataHistorical} or childs.}
}
\value{
A \link[data.table]{data.table} showing for each Bloomberg ticker, the
  Bloomberg fields for which the \code{object} provided contains data.
}
\description{
Access method for the \code{fields} slot of various S4 objects from the
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.
}
