% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{get_info}
\alias{get_info}
\alias{get_info,DataInfo-method}
\alias{get_info,DataInfo}
\title{Accesses the \code{info} slot of S4 objects of class \linkS4class{DataInfo}
  output of \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}.}
\usage{
get_info(object)

\S4method{get_info}{DataInfo}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{DataInfo} or childs.}
}
\value{
A \link[tibble]{tibble} showing qualitative data contained in the
  \code{object} provided.
}
\description{
Access method for the \code{info} slot of S4 objects of class
  \linkS4class{DataInfo} from the \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}.
}
