% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{get_name}
\alias{get_name}
\alias{get_name,AssetPricingFactor-method}
\alias{get_name,AssetPricingFactor}
\title{Accesses the \code{name} slot of S4 objects of class \linkS4class{AssetPricingFactor} output of
  \href{https://github.com/bautheac/AssetPricingFactor/}{\pkg{AssetPricingFactor}}.}
\usage{
get_name(object)

\S4method{get_name}{AssetPricingFactor}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A scalar \code{\link[base]{character}} vector containing the name of the factor.
}
\description{
Access method for the \code{name} slot of S4 objects of class
  \linkS4class{AssetPricingFactor} from the
  \href{https://github.com/bautheac/factorem/}{\pkg{factorem}}.
}
