% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{get_periods}
\alias{get_periods}
\alias{get_periods,DataHistorical-method}
\alias{get_periods,DataHistorical}
\alias{get_periods,FuturesMarket-method}
\alias{get_periods,DataHistorical,FuturesMarket}
\alias{get_periods,FuturesCFTC-method}
\alias{get_periods,DataHistorical,FuturesCFTC}
\title{Shows the time periods for which various S4 objects output of
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} contain data.}
\usage{
get_periods(object)

\S4method{get_periods}{DataHistorical}(object)

\S4method{get_periods}{FuturesMarket}(object)

\S4method{get_periods}{FuturesCFTC}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{DataHistorical} or childs.}
}
\value{
A \link[data.table]{data.table} showing for each Bloomberg ticker and
  Bloomberg field the time periods for which the \code{object} provided
  contains data.
}
\description{
Returns a data.table showing, for various S4 objects output of
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}, fo each Bloomberg
  ticker and field, the time periods for which the object contains data.
}
