% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{get_returns}
\alias{get_returns}
\alias{get_returns,AssetPricingFactor-method}
\alias{get_returns,AssetPricingFactor}
\title{Accesses the \code{returns} slot of S4 objects of class \linkS4class{AssetPricingFactor}
  output of \href{https://github.com/bautheac/AssetPricingFactor/}{\pkg{AssetPricingFactor}}.}
\usage{
get_returns(object)

\S4method{get_returns}{AssetPricingFactor}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A \code{\link[data.table]{data.table}} of factor returns historical returns.
}
\description{
Access method for the \code{returns} slot of S4 objects of class
  \linkS4class{AssetPricingFactor} from the
  \href{https://github.com/bautheac/factorem/}{\pkg{factorem}} package.
}
