% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{get_tickers}
\alias{get_tickers}
\alias{get_tickers,EquityHistorical-method}
\alias{get_tickers,EquityHistorical}
\alias{get_tickers,FundHistorical-method}
\alias{get_tickers,FundHistorical}
\title{Accesses \code{tickers} slot of various S4 objects output of
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}.}
\usage{
get_tickers(object)

\S4method{get_tickers}{EquityHistorical}(object)

\S4method{get_tickers}{FundHistorical}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{DataHistorical} or childs.}
}
\value{
A \link[data.table]{data.table} showing the Bloomberg tickers for which the
  \code{object} provided contains data.
}
\description{
Access method for the \code{tickers} slot of various S4 objects from the
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.
}
