% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{plot_performance}
\alias{plot_performance}
\alias{plot_performance,FundMarket-method}
\alias{plot_performance,DataHistorical,FundMarket}
\alias{plot_performance,AssetPricingFactor-method}
\alias{plot_performance,AssetPricingFactor}
\title{Plots market performance for various S4 objects ouput of packages that belong to the
  \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite.}
\usage{
plot_performance(object, ...)

\S4method{plot_performance}{FundMarket}(object, ticker)

\S4method{plot_performance}{AssetPricingFactor}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{DataHistorical} and childs or \linkS4class{AssetPricingFactor}.}

\item{...}{extra parameters for \code{object} related methods.}

\item{ticker}{a scalar vector. Specifies the fund Bloomberg ticker to plot performance for.}
}
\description{
Plots historical market performance inficators for various S4 objects
  from packages belonging to the \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite.
}
\examples{
\dontrun{
  library(finRes)

  # funds ####
  ## pull data from Bloomberg with pullit ####
  fund <- BBG_fund_market(tickers = "SPY US Equity",
    start = as.character(Sys.Date() - 365L), end = as.character(Sys.Date()))
  ## plot fund performance ####
  plot_performance(fund, ticker = "SPY US Equity")

  # asset pricing factors ####
  ## pull data from Bloomberg via pullit ####
  term_structure <- BBG_futures_market(type = 'term structure',
    active_contract_tickers = c("C A Comdty", "S A Comdty", "SMA Comdty", "BOA Comdty",
      "W A Comdty", "KWA Comdty", "MWA Comdty", "O A Comdty",
      "CHEA Comdty", "V6A Comdty", "FCA Comdty", "LCA Comdty", "LHA Comdty",
      "SBA Comdty", "CCA Comdty", "CTA Comdty", "KCA Comdty", "JOA Comdty", "LBA Comdty",
      "GCA Comdty", "SIA Comdty", "PAA Comdty", "PLA Comdty", "LAA Comdty", "LLA Comdty",
      "LNA Comdty", "LPA Comdty", "LTA Comdty", "LXA Comdty",
      "CLA Comdty", "COA Comdty", "DLA Comdty", "NGA Comdty", "HOA Comdty", "XBA Comdty"),
    start = as.character(Sys.Date() - (2L * 365L)), end = as.character(Sys.Date()),
    TS_positions = 1L, roll_type = "A", roll_days = 0L, roll_months = 0L, roll_adjustment = "N")
  ## construct an asset pricing factor with factorem ####
  factor <- momentum_factor(term_structure)
  ## plot factor performance ####
  plot_performance(factor)
}


}
