% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{plot_term_structure}
\alias{plot_term_structure}
\alias{plot_term_structure,FuturesTS-method}
\alias{plot_term_structure,DataHistorical,FuturesTS}
\title{Plots futures term structure over time.}
\usage{
plot_term_structure(object, ticker, frame)

\S4method{plot_term_structure}{FuturesTS}(object, ticker, frame)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesTS}.}

\item{ticker}{a scalar character vector. Active contract Bloomberg ticker to plot the term structure for.}

\item{frame}{a scalar integer vector. Animation speed parameter; the lower the faster.}
}
\description{
Plots historical futures term structure data contained in an S4 object
  of class \linkS4class{FuturesTS} from the \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.
}
\examples{
\dontrun{
  library(finRes)
  term_structure <- BBG_futures_market(type = 'term structure',
    active_contract_tickers = "C A Comdty", start = as.character(Sys.Date() - 365L),
    end = as.character(Sys.Date()), TS_positions = 1L:10L,
    roll_type = "A", roll_days = 0L, roll_months = 0L, roll_adjustment = "N")
  plot_term_structure(term_structure, ticker = "C A Comdty")
}

}
