% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{show,DataInfo-method}
\alias{show,DataInfo-method}
\alias{show,DataInfo}
\alias{show,FuturesHistorical-method}
\alias{show,FuturesHistorical}
\alias{show,EquityHistorical-method}
\alias{show,EquityHistorical}
\alias{show,FundHistorical-method}
\alias{show,FundHistorical}
\alias{show,FuturesTS-method}
\alias{show,FuturesTS}
\alias{show,FuturesCFTC-method}
\alias{show,FuturesCFTC}
\alias{show,AssetPricingFactor-method}
\alias{show,AssetPricingFactor}
\title{Show method for various S4 objects in the \href{https://bautheac.github.io/finRes/}{\pkg{finRes}}
  suite.}
\usage{
\S4method{show}{DataInfo}(object)

\S4method{show}{FuturesHistorical}(object)

\S4method{show}{EquityHistorical}(object)

\S4method{show}{FundHistorical}(object)

\S4method{show}{FuturesTS}(object)

\S4method{show}{FuturesCFTC}(object)

\S4method{show}{AssetPricingFactor}(object)
}
\arguments{
\item{object}{an S4 object fron the \href{https://bautheac.github.io/finRes/}{\pkg{finRes}}
suite.}
}
\description{
Show method for various S4 objects in the \href{https://bautheac.github.io/finRes/}{\pkg{finRes}}
  suite.
}
