% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,AssetPricingFactor-method}
\alias{summary,AssetPricingFactor}
\title{Factor performance summary}
\usage{
summary(object, leg = "factor")

\S4method{summary}{AssetPricingFactor}(object, leg = "factor")
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor}.}

\item{leg}{a character vector. 'long', 'short' and 'factor' display
the returns for the long, short and the factor itself respectively.
Defaults to 'factor'.}
}
\value{
A \code{\link[base]{data.frame}} with months as colums and years as rows.
  Column total shows the cumulated return for the corresponding year.
}
\description{
Factor performance summary method. Returns factor returns
  by month and year.
}
